requirement.txt
sanic==0.7.0
gunicorn==19.8.1
gino==0.7.3
alembic==0.9.9
psycopg2==2.7.4
psycopg2-binary==2.7.4
simple-bcrypt==0.1.0
Sanic-WTF==0.5.0
zxcvbn-python==4.4.18
PyJWT==1.3.0
aiosmtplib==1.0.2
Jinja2==2.10
sanic-jinja2==0.7.0
AoikLiveReload==0.1.0
dependency-injector==3.14.3
sanic-admin==0.0.6
secure==0.2.1
Sanic-GraphQL==1.1.0
graphene>=2.1.3



redis==2.10.5
